streamlit_app/app/data_unificator/requirements.txt
streamlit
pandas
numpy
openpyxl
pyxlsb
pyarrow
xlrd
psycopg2-binary
sqlalchemy
autogluon
autoviz
tqdm
multiprocessing
watchdog
pyyaml
xlwt
lxml
difflib
scikit-learn
unidecode
rapidfuzz
fuzzywuzzy
jsonschema
matplotlib
seaborn
